requirements.txt
streamlit
langchain
langchain-community
playwright
requests
beautifulsoup4
python-dotenv
openai
faiss-cpu
langchain_openai
